external/meraki-magic-mcp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
build==1.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
flask==3.1.2
flask-cors==6.0.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
linkify-it-py==2.0.3
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.25.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
meraki==2.0.3
multidict==6.7.0
packaging==25.0
pip==25.3
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pyproject-hooks==1.2.0
pytest==8.4.2
python-dotenv==1.2.1
python-multipart==0.0.21
pywin32==311
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
setuptools==78.1.1
sse-starlette==3.1.2
starlette==0.50.0
textual==7.0.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uc-micro-py==1.0.3
urllib3==2.6.2
uvicorn==0.40.0
werkzeug==3.1.4
wheel==0.45.1
yarl==1.22.0
